requirements_sentry.txt
# 🐍 Dependências para Projeto Python com Sentry
# ================================================

# Sentry SDK - Core
sentry-sdk[fastapi]==1.40.0

# Framework Web (exemplo com FastAPI)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de Dados (opcional)
sqlalchemy==2.0.23
alembic==1.12.1

# Utilitários
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Performance & Monitoring
psutil==5.9.6
prometheus-client==0.19.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.0